Refactor out React Final form and fix the layouts
I was enountering issues with React Final Form sharing the state between the forms, most likely because I was using it wrong. So when you switched tabs, it would lose your data.

This commit changes that so that data is shared between the inputs and you can properly configure things now

I've also fixed the layouts so they're a bit (emphasis on bit) nicer